A4: subclass the Draft_OrthoArray tool
This showcases how to add the new `Draft_OrthoArray` tool
from the Draft Workbench into Assembly4.

It imports the required modules, and then subclasses it,
and adds a bit of specific code to only be able to select
`App::Link` objects.

The original `Draft_OrthoArray` launches its own task panel
interface in order to create the array in this way.

What is missing is specific code that is able to rename
the created array, and place it in the `Model` object
of Assembly4. To do this, probably `Draft_OrthoArray`
needs to be modified to accept a final callback
from any subclass.

A very similar process as the one described in this commit
could be done to include the `Draft_PolarArray` and
`Draft_CircularArray` tools in Assembly4.