A4: set up the Draft toolbar, the Snapper, and the working plane
The `DraftToolBar` is set up by the `DraftGui` module. This toolbar
interacts with the snapping mechanism of the Draft workbench.

The snapping tools are set up by the `Snapper` class,
loaded by the `gui_snapper` module. The `Snapper` is installed
as the general `FreeCADGui.Snapper` attribute.

Snapping also depends on the working plane which is loaded
by the `WorkingPlane` module. The plane is installed as the
general `FreeCAD.DraftWorkingPlane` attribute.

These three elements are necessary so that the array GuiCommands
work correctly, as they set some callbacks that interact
with the 3D view, and expect to snap to the working plane.