Changes in the behavior of the object and fastener placement commands.
To the task dialogs 'FastenersLibs.placeFastenerUI' and 'placeLinkCmd.placeLinkUI'
the ability to select HoleAxes in 3D View has been added.

Steps for Part:
1. choose a part;
2. invoke the operation 'Move / Attach a Part in the assembly';
3. in 3D Viwer, select HoleAxes in the part to be attached,
        and in the part to which you want to attach;
        the order of selection is not important.

Steps for Fastener:
1. choose a fastener;
2. invoke the 'Edit Placement of a Fastener' operation;
3. in 3D Viwer, select the HoleAxis in the part to attach to.

PS: When invoking an operation in 3D Viwer, the entire assembly is selected,
    wait for the selection was canceled and then perform the selection of HoleAxes.

PPS: When clicking on an axis, a part or other element in 3D View can be selected,
     it is necessary to select the axis (it may take several clicks).
     The selected axis should be highlighted in corresponding list in the task dialog.

PPPS: When selection is performed there is less problems if the display mode of objects is 'Wireframe'.