Added AnimationExport-Option.
animationExporter provides a GUI and facilities to natively export the animation as gif and mp4 or frame by frame for further use with ffmpeg, giski etc.
The exporter features basic image composition functionality to support the use of background images, logo-overlay and a (quirky) artistic shadow.

animationExporter can also be useful outside the immediate asm4-variable-based animation-scenario, as it exposes a simple interface (cf. AnimationProvider.py)