super infoPart and makeBom
the infoPart all configurable with manuel field and auto filling field.
and there is a manage for the future add of new auto-field in Asm4 update.
you can use directly BoM without worrying about filling in the partInfo fields.
I write only in english and i comment the code and i clean it .
Try a commit in freeCAD before read the code in detail.