Features for ExpressionArray
Added support for lines and circles as Axes. Added support for Scaling which led to a Mirror Array replacing mirrorPart. An evaluation order solver for expressions because the builtin resolver couldn't handle sub-propertys of a placement property. A limitation that hindered a lot of creative and simpler solutions. Side effect: Twice as fast. Shortened property names for them to fit better in expressions and Property view. Using a PropertyLinkSub as Axis which can hold subelements property, As a consequence, removed AxisXYZ