Update translation script and Spanish translation
The script was updated to enable the update of the `asm4.ts` file every time
any other translation file is created or updated. This is made with the
objective of having this file updated with the most current changes.

If a translator doesn't have access to `pylupdate` command it can start
translating based on the `asm4.ts` file, it's only needed to replace

<TS version="2.1">

for

<TS version="2.1" language="es-AR" sourcelanguage="en">

using their language code (must be supported on FreeCAD).